Give splitter-lambda module its own package
